refactor: rename utils file
